fix: 修复 17 个编译错误 — Shizuku API / ApkVerifier API / 类型转换

+++ app/src/main/java/com/apkagent/apktools/advanced/ZipAlignTool.kt
@@ -87,7 +87,7 @@ object ZipAlignTool : Tool {
 
                 // 检查是否需要对齐
                 var needsAlignment = false
-                raf.seek(centralDirOffset)
+                raf.seek(centralDirOffset.toLong())
                 val entryBuf = ByteArray(46)
                 for (i in 0 until totalEntries) {
                     val entryStart = raf.filePointer
